move getTeamWorkflows API to TeamController
