+++ src/main/java/io/boomerang/controller/WorkflowsController.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
@@ -33,11 +32,6 @@ public class WorkflowsController {
   @GetMapping(value = "template")
   public List<TemplateWorkflowSummary> getTemplateWorkflows() {
     return workflowService.getTemplateWorkflows();
-  }
-  
-  @GetMapping(value = "/{teamId}")
-  public List<WorkflowSummary> getTeamWorkflows(@PathVariable String teamId) {
-    return workflowService.getWorkflowsForTeam(teamId);
   }
 
 }
